refactor(web): split sidebar drag and drop
